+++ src/app/chat/[id]/request-success/[reference-id]/page.tsx
@@ -1,4 +1,4 @@
-import RequestSuccessForm from '@/app/chat/[id]/request-success/_shared/components/request-success-form/request-success-form';
+import RequestSuccessForm from '@/app/chat/[id]/request-success/[reference-id]/_shared/components/request-success-form/request-success-form';
 import PageFrame from '@/shared/components/layout/page-frame/page-frame';
 
 const RequestSuccessPage = () => {
