--- src/app/chat/[id]/request-success/_shared/components/change-bank-bottom-sheet/change-bank-bottom-sheet.module.scss
@@ -1,13 +1,0 @@
-@use '@/styles/index' as *;
-
-.content_container {
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-  background-color: var(--grayscale-100);
-}
-
-.title {
-  @include title-md;
-  color: var(--grayscale-700, #222);
-}
